Fixed lingering dot with VanillaHUD
Fixed a compatibility issue with VanillaHUD Plus where the interaction dot would be shown at all times, even when not interacting or looking at an interaction object
Note: As a side effect of this fix, the HUD animation for the interaction completion on the interact circle for this mod may not trigger when using VanillaHUD Plus. This is purely cosmetic and does not affect gameplay or the interaction itself in any way.